+++ MyTunes.API/Controllers/ArtistsController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using MyTunes.Application.Commands.CreateArtist;
 using MyTunes.Application.Commands.DeleteArtist;
 using MyTunes.Application.Commands.UpdateArtist;
+using MyTunes.Application.Queries.ArtistExists;
 using MyTunes.Application.Queries.GetArtistById;
 using MyTunes.Application.Queries.GetArtists;
 using MyTunes.Application.ViewModels.Artist;
@@ -39,13 +40,16 @@ namespace MyTunes.API.Controllers
         }
 
         // api/artists/{id} GET
-        [HttpGet("{id}", Name = "GetArtistById")]
+        [HttpGet("{id:int}", Name = "GetArtistById")]
         public async Task<IActionResult> GetById(int id)
         {
-            // TODO: validar se existe Artist com o id informado e retornar NotFound caso contrário
-            
-            var viewModel = await _mediator.Send(new GetArtistByIdQuery(id));
-            return Ok(viewModel);
+            if (await _mediator.Send(new ArtistExistsQuery(id)))
+            {
+                var viewModel = await _mediator.Send(new GetArtistByIdQuery(id));
+                return Ok(viewModel);
+            }
+
+            return NotFound();
         }
 
         // api/artists POST
@@ -58,24 +62,30 @@ namespace MyTunes.API.Controllers
         }
 
         // api/artists/{id} PUT
-        [HttpPut("{id}", Name = "UpdateArtist")]
+        [HttpPut("{id:int}", Name = "UpdateArtist")]
         public async Task<IActionResult> Put(int id, [FromBody] UpdateArtistCommand command)
         {
-            // TODO: validar se existe Artist com o id informado e retornar NotFound caso contrário
+            if (await _mediator.Send(new ArtistExistsQuery(id)))
+            {
+                command.SetId(id);
+                _ = await _mediator.Send(command);
+                return NoContent();
+            }
 
-            command.SetId(id);
-            _ = await _mediator.Send(command);
-            return NoContent();
+            return NotFound(); 
         }
 
         // api/artists/{id} DELETE
-        [HttpDelete("{id}", Name = "DeleteArtist")]
+        [HttpDelete("{id:int}", Name = "DeleteArtist")]
         public async Task<IActionResult> Delete(int id)
         {
-            // TODO: validar se existe Artist com o id informado e retornar NotFound caso contrário
+            if (await _mediator.Send(new ArtistExistsQuery(id)))
+            {
+                _ = await _mediator.Send(new DeleteArtistCommand(id));
+                return NoContent();
+            }
 
-            _ = await _mediator.Send(new DeleteArtistCommand(id));
-            return NoContent();
+            return NotFound();
         }
     }
 }
